Moved enum styles and accent color to external files.
