+++ jmetro/src/main/java/jfxtras/styles/jmetro8/Accent.java
@@ -1,0 +1,56 @@
+package jfxtras.styles.jmetro8;
+
+public enum Accent {
+    BLUE,
+    BLUE_GRAY,
+    BRICK_RED,
+    CAMOUFLAGE,
+    CAMOUFLAGE_DESERT,
+    COOL_BLUE,
+    COOL_BLUE_BRIGHT,
+    GOLD,
+    GRAY,
+    GRAY_BROWN,
+    GRAY_DARK,
+    GREEN,
+    IRIS_PASTEL,
+    IRIS_SPRING,
+    LIDDY_GREEN,
+    MEADOW_GREEN,
+    METAL_BLUE,
+    MINT_DARK,
+    MINT_LIGHT,
+    MOD_RED,
+    MOSS,
+    NAVY_BLUE,
+    ORANGE_BRIGHT,
+    ORANGE_DARK,
+    ORCHID,
+    ORCHID_LIGHT,
+    OVERCAST,
+    PALE_MOSS,
+    PALE_RED,
+    PALE_RUST,
+    PLUM,
+    PLUM_LIGHT,
+    PURPLE_SHADOW,
+    PURPLE_SHADOW_DARK,
+    RED,
+    ROSE,
+    ROSE_BRIGHT,
+    RUST,
+    SAGE,
+    SEAFOAM,
+    SEAFOAM_TEAL,
+    SPORT_GREEN,
+    STEEL_BLUE,
+    STORM,
+    TURF_GREEN,
+    VIOLET_RED,
+    VIOLET_RED_LIGHT,
+    YELLOW_GOLD;
+
+    protected String getColorName() {
+        return "JMETRO_" + name();
+    }
+}
