+++ jmetro/src/main/java/jfxtras/styles/jmetro8/JMetro.java
@@ -35,81 +35,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 public class JMetro {
-    public enum Style {
-        LIGHT,
-        DARK;
-
-        private String getStyleSheetFileName() {
-            String stylesheet = null;
-
-            switch (this) {
-                case LIGHT:
-                    stylesheet = "JMetroLightTheme.css";
-                    break;
-                case DARK:
-                    stylesheet = "JMetroDarkTheme.css";
-                    break;
-            }
-
-            return stylesheet;
-        }
-    }
-
-    public enum Accent {
-        BLUE,
-        BLUE_GRAY,
-        BRICK_RED,
-        CAMOUFLAGE,
-        CAMOUFLAGE_DESERT,
-        COOL_BLUE_BRIGHT,
-        COOL_BLUE,
-        GOLD,
-        GRAY_BROWN,
-        GRAY,
-        GRAY_DARK,
-        GREEN,
-        IRIS_PASTEL,
-        IRIS_SPRING,
-        LIDDY_GREEN,
-        MEADOW_GREEN,
-        METAL_BLUE,
-        MINT_DARK,
-        MINT_LIGHT,
-        MOD_RED,
-        MOSS,
-        NAVY_BLUE,
-        ORANGE_BRIGHT,
-        ORANGE_DARK,
-        ORCHID,
-        ORCHID_LIGHT,
-        OVERCAST,
-        PALE_MOSS,
-        PALE_RED,
-        PALE_RUST,
-        PLUM,
-        PLUM_LIGHT,
-        PURPLE_SHADOW,
-        PURPLE_SHADOW_DARK,
-        RED,
-        ROSE_BRIGHT,
-        ROSE,
-        RUST,
-        SAGE,
-        SEAFOAM,
-        SEAFOAM_TEAL,
-        SPORT_GREEN,
-        STEEL_BLUE,
-        STORM,
-        TURF_GREEN,
-        VIOLET_RED,
-        VIOLET_RED_LIGHT,
-        YELLOW_GOLD;
-
-        private String getColorName() {
-            return "JMETRO_" + name();
-        }
-    }
-
     private Style style;
     private Accent accent;
     private Color accentColor;
